Compilation with gcc supported (not tested)
Makefile was corrected and improved to be more smart in compilation selection.
Deletes some unused variables in the inits.
CHanged solver variable string by solvername, to avoid mistakes.
Added the results dir to save option.
Corrected a bug in the file input parser.
Deleted the warning relates with wrong orders in constructors.