First commit making the simplification for the master slave system.
The slave clase was deleted in favor of base_calculator, that contains now all the methods for paralelization, mpi and threads.
Auxiliary was also deleted because it is not used and any of the functions called.
The filename was redefined. for standar save.
The macros in the header were inserted in the base_calculator header.
Makefile was also modified.